DW-1680 release max sndbuf-size

+++ drbd-headers/drbd_transport.h
@@ -227,7 +227,7 @@ struct drbd_transport_ops {
 	int (*add_path)(struct drbd_transport *, struct drbd_path *path);
 	int (*remove_path)(struct drbd_transport *, struct drbd_path *path);
 #ifdef _WIN32_SEND_BUFFING 
-	bool (*start_send_buffring)(struct drbd_transport *, int size);
+	bool (*start_send_buffring)(struct drbd_transport *, signed long long size);
 	void (*stop_send_buffring)(struct drbd_transport *);
 #endif
 };
